refactor(project): organize code into modules, use a common constant for the app version
