+++ lib/git-releaselog/version.rb
@@ -1,0 +1,5 @@
+module Releaselog
+    VERSION     = "0.6.1"
+    SUMMARY     = "Generate a releaselog from a git repository"
+    DESCRIPTION = "Write your releaselog as part of your usual commit messages. This tool generates a useful releaselog from marked lines in your git commit messages"
+end
